Renamed files of the parser settings and manager and fixed the entry for the stress
