+++ aiida_vasp/parsers/tests/test_parser_settings.py
@@ -1,7 +1,7 @@
 # pylint: disable=unused-import,too-few-public-methods,missing-docstring,no-self-use,no-member
 """Test the Settings utils Module."""
 
-from aiida_vasp.parsers.parser_settings import ParserSettings
+from aiida_vasp.parsers.settings import ParserSettings
 from aiida_vasp.parsers.vasp import DEFAULT_OPTIONS
 
 SETTINGS = {
